Fix static interactions
